+++ slash/react/src/Form/File/FileInput.mdx
@@ -1,5 +1,5 @@
 import { Meta, Canvas } from "@storybook/addon-docs";
-import { ArgsTable } from "@storybook/addon-docs";
+import { Controls } from "@storybook/blocks";
 import * as FileInputStories from "./FileInput.stories";
 
 # File Input
@@ -17,7 +17,7 @@ import * as FileInputStories from "./FileInput.stories";
 ### Import
 
 ```javascript
-import { FileInput } from "@axa-fr/design-system/agent";
+import { FileInput } from "@axa-fr/design-system-slash-react/agent";
 ```
 
 ### Use
@@ -41,7 +41,7 @@ export default FileInputReturn;
 
 <Canvas of={FileInputStories.FileInputStory} />
 
-<ArgsTable story="File" />
+<Controls story="File" />
 
 ### FileInput Required
 
@@ -58,7 +58,7 @@ to the **FileInput** component
 ### Import
 
 ```javascript
-import { File } from "@axa-fr/design-system/agent";
+import { File } from "@axa-fr/design-system-slash-react/agent";
 ```
 
 ### Use
@@ -88,7 +88,11 @@ export default fileReturn;
 ### Import
 
 ```javascript
-import { File, FileTable, HelpMessage } from "@axa-fr/design-system/agent";
+import {
+  File,
+  FileTable,
+  HelpMessage,
+} from "@axa-fr/design-system-slash-react/agent";
 ```
 
 ### Use
@@ -155,7 +159,11 @@ export default FileWithValues;
 ### Import
 
 ```javascript
-import { File, FileTable, HelpMessage } from "@axa-fr/design-system/agent";
+import {
+  File,
+  FileTable,
+  HelpMessage,
+} from "@axa-fr/design-system-slash-react/agent";
 ```
 
 ### Use
